feat: refactor advanced resources page
