+++ src/components/NodeCard.tsx
@@ -2,7 +2,7 @@ import { Handle, Position } from "reactflow";
 import "reactflow/dist/style.css";
 import Star from "@/components/icons/Star";
 import classNames from "classnames";
-import { getColorByAccess } from "@/utils/Helper";
+import { getColorByAccess } from "@/utils/helpers";
 
 export function NodeCard({ data, id }: any) {
   id = data.Key;
